+++ tsup.config.ts
@@ -16,6 +16,7 @@ export default defineConfig({
     'src/brand.ts',
     'src/locale/index.ts',
     'src/utils/*.ts',
+    'src/hooks/*.ts',
     'src/components/!(*.stories).tsx',
     'src/presets/elalba/defaults.ts',
   ],